Back/requirements.txt
asgiref==3.5.2
gunicorn==20.1.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
Pillow==9.2.0
# added 
dj-database-url==1.0.0
Django==3.2.16
django-shortcuts==1.6
psycopg2-binary==2.9.5
tzdata==2022.1